feat(taskboard): phase 4 (final) — ingestion.{store,models} 物理删 + cascade + iteration close-out
Yi #133 + HUMAN #152 + Orchestrator #158 lock Phase 4 范围 (~1500 行原 lock,
不 inflate). 整个 frago Msg/Task 重构迭代收官 PR.

## 4 项 deliverable (Yi #133 lock)

1. **ingestion/store.py 物理删除** ✓
   - 旧路径 src/frago/server/services/ingestion/store.py 已 git rm
   - TaskStore 类迁入 taskboard/legacy_store.py (clean import path, board 单一持
     久化层, IngestedTask 持有执行上下文 sidecar 字段)
   - 9 caller files imports 从 ingestion.store → taskboard.legacy_store cascade

2. **ingestion/models.py 物理删除** ✓
   - 整文件 git rm (114 行)
   - IngestedTask + SubTask + TaskStatus 三类迁入 taskboard/models.py
   - taskboard/__init__.py 增加 TaskStatus 导出
   - 9 caller files imports 从 ingestion.models → taskboard.models cascade

3. **frontend API contract 维持** ✓
   - /pa/tasks endpoint 不破坏 wire 兼容 (返回 IngestedTask asdict 形态)
   - frontend TypeScript IngestedTask 类型 (client/src/types/*) 不动

4. **rename test_vacuum_fold.py + .gitignore allowlist 通配** ✓ (Kai #110 承诺 close-out)
   - tests/server/test_vacuum_fold.py → tests/server/test_taskboard_vacuum_fold.py
   - .gitignore 移除 `!tests/server/test_vacuum_fold.py` 行 (回归 test_taskboard_*.py 通配)

## 验收

- **56 passed + 0 skipped + 0 failed** (含 9 测试模块 regression-free)
- **ruff All checks passed** (touched files)
- **grep verify 0 残留 import**:
  - `from frago.server.services.ingestion.{store,models}` → 0 matches
  - `import frago.server.services.ingestion.{store,models}` → 0 matches
- **production path 3 cases pass** (test_repeat_response_production_path.py 不退化)

## ingestion/ 目录现状

仅保留 __init__.py + scheduler.py (active scheduler 模块, 仍是消息入口).
store.py + models.py 已物理删, legacy 类全部迁入 taskboard layer.

## Iteration close-out 声明

整个 frago Msg/Task 重构迭代 (始于 spec 20260512-msg-task-board-redesign)
通过 8 个 PR 完成:
- PR #67 Phase 0 (timeline.jsonl persistence + 4-applier skeleton)
- PR #68 Phase 1A (4 Applier interfaces + ingest_scheduled)
- PR #69 Phase 1B-1 (msg/task close methods)
- PR #70 Phase 1B-2a (production-path repeat-response root-cause fix)
- PR #71 Phase 2 (vacuum + fold + CLI)
- PR #72 Phase 1B-2b stage 1 (thread_service deletion + 7 board methods)
- PR #73 Phase 3 (PA cleanup full rewrite)
- **本 PR Phase 4 (ingestion 物理删 + cascade) — 收官**
