+++ .gitignore
@@ -26,7 +26,6 @@ tests/*
 tests/server/*
 !tests/server/test_taskboard_*.py
 !tests/server/test_repeat_response_production_path.py
-!tests/server/test_vacuum_fold.py
 tests/results/
 tests/*.log
 coverage/
