+++ src/frago/server/services/executor.py
@@ -19,9 +19,9 @@ from datetime import datetime
 from pathlib import Path
 from typing import Any
 
-from frago.server.services.ingestion.models import IngestedTask, TaskStatus
-from frago.server.services.ingestion.store import TaskStore
 from frago.server.services.resume_inbox import ResumeInbox
+from frago.server.services.taskboard.legacy_store import TaskStore
+from frago.server.services.taskboard.models import IngestedTask, TaskStatus
 
 logger = logging.getLogger(__name__)
